fix bug

+++ mit_d3m/__init__.py
@@ -52,19 +52,53 @@ def get_dataset_s3_key(dataset):
 
 
 def download_dataset(bucket, key, filename):
+    """Download dataset from s3://bucket/key to filename"""
     print("Downloading dataset from s3://{bucket}".format(bucket=bucket))
     client = get_client()
     client.download_file(Bucket=bucket, Key=key, Filename=filename)
 
 
 def extract_dataset(src, dst):
+    """Extract tarfile at src to within dst
+
+    Removes a directory at dst if such a directory exists. Note that the dataset archive
+    generally contains members with a common directory prefix, i.e. all the files for a dataset
+    named ``185_baseball`` have the common prefix ``185_baseball``. Thus after extraction, all
+    files will be underneath ``dst/185_baseball`` or similar.
+
+    Args:
+        src (path-like): path to tarfile, which should be gzipped
+        dst (path-like): path to destination directory
+
+    Raises:
+        ValueError: the source path is not a valid tarfile
+    """
     print("Extracting {}".format(src))
-    shutil.rmtree(dst, ignore_errors=True)
+
+    if not (os.path.exists(src) and tarfile.is_tarfile(src)):
+        raise ValueError('Invalid source path: {src}'.format(src=src))
+
+    if os.path.exists(dst) and os.path.isdir(dst):
+        shutil.rmtree(dst, ignore_errors=True)
+
+    os.makedirs(dst)
     with tarfile.open(src, 'r:gz') as tf:
         tf.extractall(dst)
 
 
 def load_d3mds(dataset, root=DATA_PATH, force_download=False):
+    """Load dataset into D3MDS format, as necessary downloading tarfile from S3 and extracting
+
+    Args:
+        dataset (str): dataset identifier
+        root (path-like, optional): root directory to store tarfiles and extracted datasets.
+            Defaults to './data/'.
+        force_download (boolean, optional): download the tarfile even if it already exists,
+            also causing the files to be re-extracted. Defaults to False.
+
+    Returns:
+        mit_d3m.dataset.D3MDS
+    """
     read_only = root != DATA_PATH
 
     if not read_only and not os.path.exists(root):
@@ -74,7 +108,7 @@ def load_d3mds(dataset, root=DATA_PATH, force_download=False):
         dataset = dataset[:len(DATASET_EXTRA_SUFFIX)]
 
     dataset_dir = get_dataset_dir(root, dataset)
-    dataset_tarfile = get_dataset_tarfile_path(dataset_dir, dataset)
+    dataset_tarfile = get_dataset_tarfile_path(root, dataset)
     dataset_key = get_dataset_s3_key(dataset)
 
     requires_download = force_download or not os.path.exists(dataset_tarfile)
@@ -87,7 +121,7 @@ def load_d3mds(dataset, root=DATA_PATH, force_download=False):
     if not read_only and requires_extraction:
         extract_dataset(dataset_tarfile, dataset_dir)
 
-    phase_root = os.path.join(dataset_dir, 'TRAIN')
+    phase_root = os.path.join(dataset_dir, dataset, 'TRAIN')
     dataset_path = os.path.join(phase_root, 'dataset_TRAIN')
     problem_path = os.path.join(phase_root, 'problem_TRAIN')
 
@@ -95,8 +129,19 @@ def load_d3mds(dataset, root=DATA_PATH, force_download=False):
 
 
 def load_dataset(dataset, root=DATA_PATH, force_download=False):
+    """Load dataset, as necessary downloading tarfile from S3 and extracting
 
-    d3mds = load_d3mds(dataset, root, force_download=force_download)
+    Args:
+        dataset (str): dataset identifier
+        root (path-like, optional): root directory to store tarfiles and extracted datasets.
+            Defaults to './data/'.
+        force_download (boolean, optional): download the tarfile even if it already exists,
+            also causing the files to be re-extracted. Defaults to False.
+
+    Returns:
+        mit_d3m.loaders.Dataset
+    """
+    d3mds = load_d3mds(dataset, root=root, force_download=force_download)
 
     loader = get_loader(
         d3mds.get_data_modality(),
